Added Conversion $ Parsing to separate classes (removed from GeneBankCreateBTree). Need to update search to use ParseFile class as well. Working ParseTest that appears to be parsing correct information and ignoring 'n'. Will continue past // if more ORIGIN is found.
