+++ GeneBankCreateBTree.java
@@ -128,108 +128,9 @@ public class GeneBankCreateBTree {
             }
         }
 
+        File gbk_file = new File(args[2]);
 
-
-        try {
-            File gbk_file = new File(args[2]);
-
-            // Creates the File/Buffered Readers to process gbk file
-            FileReader readFile = new FileReader(gbk_file);
-            BufferedReader readBuffer = new BufferedReader(readFile);
-            // Variable that holds the current line of text being processed
-            String currentLine;
-            StringBuilder currentSequence = new StringBuilder();
-            String sequence;
-
-            char currentChar;
-            int currentLinePosition = 0;
-            int currentSequencePosition = 0;
-
-            boolean parseRun = false;
-            boolean parseComplete = false;
-
-
-            // Reads through gbk file to parse out the sequences
-            while ((currentLine = readBuffer.readLine()) != null) {  // trim? toLower/Upper?
-                if (currentLine.startsWith("ORIGIN")) {
-                    parseRun = true;
-                }
-                if (parseRun) {
-                    if (currentLine.startsWith("//")) {
-                        parseRun = false;
-                        parseComplete = true;
-                        //currentSequencePosition = 0;
-                        break;
-                    } else {
-
-                        while (currentLinePosition < currentLine.length()) {
-                            currentChar = currentLine.charAt(currentLinePosition); // toLower/Upper?
-                            currentLinePosition++;
-
-                            // Inserts character into sequence
-                            switch (currentChar) {
-                                case 'a':
-                                    currentSequence.append(currentChar);
-                                    break;
-                                case 't':
-                                    currentSequence.append(currentChar);
-                                    break;
-                                case 'c':
-                                    currentSequence.append(currentChar);
-                                    break;
-                                case 'g':
-                                    currentSequence.append(currentChar);
-                                    break;
-                                case 'n':
-                                    currentSequence.append(currentChar);
-                                    break;
-//                                    // Ends subsequence and resets parsing for a new subsequence (DONT need in parsing sequence)
-//                                    currentSequencePosition = 0;
-//                                    currentSequence.delete(0, currentSequence.length());
-//                                    continue;
-                                default:
-                                    // Skips white space and number characters
-                                    continue;
-                            }
-                        }
-                    }
-                }
-                sequence = currentSequence.toString();
-                long key = 0;
-                int counter = 0;
-                while (currentSequencePosition < sequence.length()) {
-                    StringBuilder currentSubseq = new StringBuilder();
-
-                    while (counter < subsequenceLength) {
-                        currentSubseq.append(sequence.charAt(currentSequencePosition));
-
-                        counter++;
-                        currentSequencePosition++;
-                    }
-
-                    String parsedSubseq = currentSubseq.toString();
-                    counter = 0;
-
-                    key = subsequenceToLong(parsedSubseq, subsequenceLength);
-                    // Insert subsequence key into tree
-//                    myBTree.insert(key);
-                }
-            }
-            // Closes the File/Buffer Readers
-            readBuffer.close();
-            readFile.close();
-        }
-        // Catches exception if file not found
-        catch(FileNotFoundException e) {
-            System.out.println("ERROR: Cannot open file : " + e.getMessage() + "\n\n");
-            System.out.println("Make sure that the gbk file is in the same folder as the HashTest and other java files for this project.");
-            System.exit(0);
-        }
-        // Catches exception if error closing readers
-        catch (IOException e) {
-            System.out.println("ERROR: When closing Buffer/File : " + e.getMessage() + "\n\n");
-            System.exit(0);
-        }
+        ParseFile.parseGbk(gbk_file, subsequenceLength);
 
         // Checks debug level and prints to file accordingly
         try {
@@ -268,75 +169,5 @@ public class GeneBankCreateBTree {
         //System.out.println("Usage: java GeneBankCreateBTree <degree> <gbk file> <sequence length> [<debug level>]\n");
         System.out.println("Usage: java GeneBankCreateBTree <cache> <degree> <gbk file> <sequence length> <cache size> [<debug level>]\n");   // cache optional?
         System.exit(0);
-    }
-
-//    static long getNextSubSequence(String seq, int k, int startIndex) {
-//        long key = 0;
-//
-//        int seqPosition = startIndex;
-//        int counter = 0;
-//        StringBuilder currentSubseq = new StringBuilder();
-//
-//
-//        // Parses out subsequences until end of sequence; if end of sequence or error return -1
-//        while (seqPosition < seq.length()){
-//            while (counter < k) {
-//                currentSubseq.append(seq.charAt(seqPosition));
-//
-//                seqPosition++;
-//                counter++;
-//            }
-//            counter = 0;
-//        }
-//
-//
-//
-//        return key;
-//    }
-
-    static long subsequenceToLong(String subseq, int k) {
-        long key = 0x00;
-        long temp = 0;
-        int shift;
-
-        for (int i = 0; i < k; i++) {
-            shift = 2*i;
-            switch (subseq.charAt(i)) {
-                case 'a':
-                    temp = BASE_A << shift;
-                    break;
-                case 't':
-                    temp = BASE_T << shift;
-                    break;
-                case 'c':
-                    temp = BASE_C << shift;
-                    break;
-                case 'g':
-                    temp = BASE_G << shift;
-                    break;
-            }
-            key = key | temp;
-        }
-        return key;
-    }
-
-    static String longToSubsequence(long key, int k) {
-        StringBuilder result = new StringBuilder();
-        int shift;
-        long temp;
-        char gene;
-
-        for (int i = 0; i < k; i++) {
-            shift = 2*i;
-            temp = key;
-
-            temp = temp >> shift;
-            temp = temp & 0b11L;
-
-            gene = geneMap[(int)temp];
-            result.append(gene);
-        }
-        String subsequence = result.toString();
-        return subsequence;
     }
 }
